lipsum: generate random lorem ipsum text
The function now generates a random string with a given number of
words. The string will start with "Lorem ipsum", but will eventually
become random since the next word is based not only on the traditional
lorem ipsum text, but on the full first book in Cicero's work De
finibus bonorum et malorum ("On the ends of good and evil").