MarkovChain: add iter and iter_from methods
These methods return an iterator which will generate a neverending
chain of words. The iterator can be used by itself, or it can be used
indirectly by the generate and generate_from methods.

The new iterator works slightly differently from the old code in
corner cases where the iteration ends up in a state from which there
is no next state in the Markov chain. Like before, a new state is
picked at random, but this now happens one word earlier than before.
This should not make a big difference outside of artificially small
test cases.