MarkovChain: add rng struct field
The field carries the random number generator used by the chain. By
initializing the field with a seeded RNG, the output can be made
deterministic, which in turn can be important if the output needs to
be reproducible.

This makes generating the Markov chain about 2% faster, perhaps due to
fewer calls to thread_rng:

 name                      old ns/iter  new ns/iter  diff ns/iter  diff %
 generate_lorem_ipsum_100  47,295       46,406               -889  -1.88%
 generate_lorem_ipsum_200  70,367       68,563             -1,804  -2.56%
 learn_liber_primus        1,727,529    1,707,559         -19,970  -1.16%
 learn_lorem_ipsum         12,296       12,367                 71   0.58%