MarkovChain: use Rng trait for random number generator
Before the field was restricted to the thread local RNG. It can now be
any kind of random number generator.

This required an unexpected change to how we pick random elements: the
Rng::choose method requires that Self is Sized, which means that the
method cannot be called on a trait object. To get around this, the
choose method has been reimplemented in terms of the next_f32 method,
which has no such requirements.

The new choose function is around 6-7 faster than the old method:

 name                      old ns/iter  new ns/iter  diff ns/iter  diff %
 generate_lorem_ipsum_100  46,171       43,257             -2,914  -6.31%
 generate_lorem_ipsum_200  68,682       63,692             -4,990  -7.27%
 learn_liber_primus        1,702,205    1,683,893         -18,312  -1.08%
 learn_lorem_ipsum         12,249       12,172                -77  -0.63%