MarkovChain: ensure deterministic output
Before, the Markov chain iterators would generate the list of valid
keys on demand. Because the hash map is unordered, the keys would be
generated in a random order. This in turn meant that it wasn't
possible to generate deterministic Markov chain output.

We now cache the keys after learning from a new text. The keys are
kept sorted and the iterators get a reference to this sorted list of
keys. This means fewer allocations when iterating and the benchmark
time for generating output has dropped by about 50%:

 name                      old ns/iter  new ns/iter  diff ns/iter   diff %
 generate_lorem_ipsum_100  43,325       17,969            -25,356  -58.53%
 generate_lorem_ipsum_200  63,043       37,040            -26,003  -41.25%

On the other hand, up-front cost of learning text has increased:

 name                      old ns/iter  new ns/iter  diff ns/iter   diff %
 learn_liber_primus        1,714,541    3,999,904       2,285,363  133.29%
 learn_lorem_ipsum         12,262       21,253              8,991   73.32%

This seems okay since the Markov chain only needs to be trained once
after which it can be used many times. Furthermore, it still only
takes 4 ms to train the chain on the 48 KB Liber Primus text.