MarkovChain: make map private and add words method
The new MarkovChain::words method can be used to find the words
following a given bigram.

We will be adding more fields to MarkovChain soon and the map field
can then no longer be public. The new method solves this by giving
read-only access to the underlying HashMap, while hiding the concrete
implementation of the Markov chain.