Remove outdated Rust 1.31 badge
Back in 2020, I gave up on trying to keep the library compatible with
old Rust releases. This was with the 0.7.0 release which gave up
compatibility with Rust 1.31 (the first version of the 2018 edition).

Since then, the releases have roughly tracked the latest stable
release of Rust, with no particular urgency to bump this of course.

This is related to #83 where @WorldSEnder noticed that Lipsum no
longer compiles on Rust 1.56. This was easy to fix (#84), but we
should still remove the misleading badge.