Test with Rust 1.56 in CI
This is the first version to support the 2021 edition. I tried
tracking the minimum supported Rust version (MSRV) back in 2020, but
this turned out to be infeasible because of frequent bumps to the MSRV
in our dependencies (see 9d622cc). I hope the update pace has cooled
enough to make the feasible again.

I plan to (conservatively) bump this version up as needed going
forward. This will of course announcing it in the release notes of new
releases.

See #83 for the motivation.